Starting to carve up the graphql entities
