+++ orders_api/src/models/mod.rs
@@ -1,0 +1,3 @@
+mod burger;
+
+pub use burger::Burger;
